pyproject.toml
[project]
name = "noc-managers"
version = "0.2.0"
description = "AI-powered NOC management via SolarWinds, vCenter, and LangChain"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.42.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-mcp-adapters>=0.0.9",
    "mcp>=1.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "openai>=1.60.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/noc_managers"]

[project.scripts]
# Kept for backward compatibility
noc-app = "noc_managers.app:main"
noc-server = "noc_managers.mcp_server:main"
# Granular entry points for each MCP server
noc-server-solarwinds = "noc_managers.mcp_servers.solarwinds:main"
noc-server-vcenter = "noc_managers.mcp_servers.vcenter:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
